Fixed roll action for new input system
Modified roll penalty to add extra drag instead of a force.

Removed no-roll constraint on calls to MoveTowardCursor, ApplyDrag and
PerformActions. Instead, the constraint is put on handling of input. This
way, the player continues to move in the direction of the last input
once a roll is actuated.

fixes #3